feat(uncommit changes warning): slide in when
