+++ src/components/ChatLinks/UncommittedChangesWarning.module.css
@@ -1,0 +1,14 @@
+.changes_warning {
+  overflow-wrap: anywhere;
+  position: absolute;
+  animation: rollin 0.2s ease-out;
+}
+
+@keyframes rollin {
+  0% {
+    transform: translateY(-100%);
+  }
+  100% {
+    transform: translateY(0);
+  }
+}
